reformat and login added
