+++ src/components/Navbar.js
@@ -1,7 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import './Navbar.css'; // Import the CSS
-
+import "../css/Navbar.css"; // Import the CSS
 
 function Navbar() {
   return (
