+++ src/css/LoginPage.css
@@ -1,0 +1,113 @@
+/* HomePage.css */
+
+/* Main container for the homepage */
+.homepage {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  min-height: 100vh;
+  background: linear-gradient(135deg, #6a11cb 0%, #2575fc 100%);
+  padding: 20px;
+}
+
+/* Header section styling */
+.homepage__header h1 {
+  font-size: 3rem;
+  color: #ffffff;
+  margin-bottom: 30px;
+  text-align: center;
+}
+
+/* Actions container that holds the signup button and login form */
+.homepage__actions {
+  background-color: rgba(255, 255, 255, 0.85);
+  padding: 40px;
+  border-radius: 15px;
+  box-shadow: 0 8px 16px rgba(0, 0, 0, 0.25);
+  max-width: 400px;
+  width: 100%;
+}
+
+/* Signup button styling */
+.homepage__signup-btn {
+  display: block;
+  width: 100%;
+  background-color: #4caf50;
+  color: #ffffff;
+  border: none;
+  padding: 12px;
+  border-radius: 25px;
+  font-size: 1.2rem;
+  margin-bottom: 20px;
+  cursor: pointer;
+  transition: background-color 0.3s ease;
+}
+
+.homepage__signup-btn:hover {
+  background-color: #45a049;
+}
+
+/* Login form styling */
+.homepage__login-form {
+  display: flex;
+  flex-direction: column;
+  align-items: stretch;
+}
+
+/* Label for form inputs */
+.homepage__label {
+  font-size: 1rem;
+  color: #333;
+  margin-bottom: 5px;
+}
+
+/* Input fields styling */
+.homepage__input {
+  padding: 10px;
+  border: 2px solid #ddd;
+  border-radius: 5px;
+  font-size: 1rem;
+  margin-bottom: 15px;
+  transition: border-color 0.3s ease;
+}
+
+.homepage__input:focus {
+  border-color: #2575fc;
+  outline: none;
+}
+
+/* Error message styling */
+.homepage__error {
+  color: #ff4d4f;
+  font-size: 0.9rem;
+  margin-bottom: 15px;
+  text-align: center;
+}
+
+/* Login button styling */
+.homepage__login-btn {
+  background-color: #2575fc;
+  color: #fff;
+  border: none;
+  padding: 12px;
+  border-radius: 25px;
+  font-size: 1.2rem;
+  cursor: pointer;
+  transition: background-color 0.3s ease;
+}
+
+.homepage__login-btn:hover {
+  background-color: #1e63d4;
+}
+
+/* Responsive adjustments for smaller screens */
+@media (max-width: 600px) {
+  .homepage__header h1 {
+    font-size: 2.5rem;
+  }
+
+  .homepage__actions {
+    padding: 20px;
+  }
+}
